requirements.txt
rasa[convert]~=1.10.0
pytablewriter
pandas
pyyaml
django-environ
